experiment(ntuple-scale): preregister the fresh-block replication at larger scale
Experiment EX-20260906-ntuple-scale-replication-wide-plateau-f627f07a
tests the same theory on three never-read seed blocks with a wider
network and no fixed move budget: the first experiment's selected layout
(rows, columns, 2x3 and 3x2 windows, every family conditioned on the
rise clock) plus 2x4 and 4x2 window families of 10^8 patterns each, 5.8e9
table entries in all, trained by the same temporal-coherence TD until the
paired margin on a 256-game training-role validation block stops
improving (window-of-four plateau rule, from the eighth point), then
screened once on 512 held-out games beside the first experiment's frozen
tables, the identical fair-leaf search, and the depth-4 reference.

Crate changes: win24 and win42 families with pattern indices straight
from four- and two-nibble chunks of the packed column words; 64-bit
feature indices (the wide layout exceeds 2^32 entries); the coherence
accumulators are allocated zeroed so their pages become resident only
when touched; the rise phase is clamped for terminal states, which the
CHECK gate perturbs (the search never values them); the trainer gains
the plateau rule, a STOP file, a checkpoint cadence and stop.json. The
successor driver scripts/pipeline2.sh hard-codes the new leases;
pipeline.sh, the first experiment's frozen driver, is untouched.
analyze.py and the web snapshot understand both run shapes.

CHECK gates (19, including feature-index references for the new
families) pass on the probe block; the smoke run on the probe block
measured 1.27 million training moves per second on 32 threads with a
42.7 GB resident set after 40 million moves. Three leases are reserved,
disjoint from every block the first experiment read.

Theory-ID: TH-20260905-ntuple-line-tuples-tc-td-leaf-bcb25133
Experiment-ID: EX-20260906-ntuple-scale-replication-wide-plateau-f627f07a
Run-ID: none
Result-ID: none
Contribution-ID: CT-20260906T014308Z-21df88ae
Evidence-Change: proposed-to-preregistered

+++ approaches/ntuple-rl/ntuple-scale/src/game.rs
@@ -44,7 +44,7 @@ impl<L: drop7_rs::search::Leaf> Player for SearchPlayer<L> {
 pub struct DirectPlayer {
     pub model: Arc<Model>,
     pub params: PolicyParams,
-    scratch: [u32; MAX_ACTIVE],
+    scratch: [u64; MAX_ACTIVE],
 }
 
 impl Player for DirectPlayer {
@@ -93,7 +93,7 @@ impl Arm {
             Arm::Direct(model) => Box::new(DirectPlayer {
                 model: model.clone(),
                 params: PolicyParams::default(),
-                scratch: [0u32; MAX_ACTIVE],
+                scratch: [0u64; MAX_ACTIVE],
             }),
         }
     }
